pipeline/requirements.txt
audioread==2.1.6
backports.functools-lru-cache==1.5
cachetools==2.1.0
certifi==2018.8.24
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
eyeD3==0.8.7
funcsigs==1.0.2
futures==3.2.0
google-api-core==1.4.0
google-auth==1.5.1
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
h5py==2.8.0
hmmlearn==0.2.0
idna==2.7
joblib==0.12.5
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.24.0
matplotlib==2.2.3
numba==0.39.0
numpy==1.15.1
pathlib==1.0.1
pocketsphinx==0.1.15
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pyAudioAnalysis==0.2.5
pydub==0.23.0
pyOSC==0.3.5b5294
pyparsing==2.2.1
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
PyYAML==3.13
requests==2.19.1
resampy==0.2.1
rsa==4.0
scikit-learn==0.19.2
scipy==1.1.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.5.2
Theano==1.0.2
urllib3==1.23
wavio==0.0.4
